requirements.txt
# MedDocReader - Medical Document Processing System
# Core dependencies

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Azure services
azure-ai-formrecognizer==3.3.2
azure-storage-blob==12.19.0
azure-core==1.29.5

# NLP processing
spacy==3.7.2

# Data processing
pandas==2.1.3
numpy==1.25.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Logging and monitoring
structlog==23.2.0

